refactor(cursor): split image protobuf encoding
